Rename onboarding to intro, separate onboarding and city selector screens
