+++ kmp/features/eventReporting/src/commonMain/kotlin/com/egoriku/grodnoroads/eventreporting/screen/ui/foundation/SelectableOptions.kt
@@ -1,6 +1,7 @@
 package com.egoriku.grodnoroads.eventreporting.screen.ui.foundation
 
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.KeyboardActions
@@ -29,7 +30,7 @@ import com.egoriku.grodnoroads.shared.models.reporting.ReportParams
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-internal fun SelectableOptions(
+internal fun ColumnScope.SelectableOptions(
     reportType: Reporting.ReportType,
     onReportParamsChange: (ReportParams) -> Unit
 ) {
