+++ kmp/features/intro/build.gradle.kts
@@ -1,0 +1,35 @@
+import com.egoriku.grodnoroads.extension.applyTargets
+import com.egoriku.grodnoroads.extension.commonDependencies
+
+plugins {
+    alias(libs.plugins.grodnoroads.kmp.library)
+    alias(libs.plugins.grodnoroads.kmp.compose)
+    alias(libs.plugins.kotlin.serialization)
+}
+
+android {
+    namespace = "com.egoriku.grodnoroads.intro"
+}
+
+kotlin {
+    applyTargets()
+
+    sourceSets {
+        commonDependencies {
+            implementation(projects.kmp.compose.commonUi)
+            implementation(projects.kmp.compose.foundation.icons)
+            implementation(projects.kmp.compose.foundation.uikit)
+            implementation(projects.kmp.compose.resources)
+            implementation(projects.kmp.shared.persistent)
+            implementation(projects.kmp.libraries.logger)
+            implementation(projects.kmp.libraries.extensions)
+
+            implementation(libs.decompose)
+            implementation(libs.decompose.compose)
+            implementation(libs.koin.core)
+            implementation(libs.kotlin.collections)
+
+            implementation(libs.bundles.mvikotlin)
+        }
+    }
+}
