+++ kmp/features/intro/src/commonMain/kotlin/com/egoriku/grodnoroads/cityselector/domain/component/CitySelectorComponent.kt
@@ -1,0 +1,23 @@
+package com.egoriku.grodnoroads.cityselector.domain.component
+
+import androidx.compose.runtime.Stable
+import com.egoriku.grodnoroads.cityselector.domain.store.CitySelectorStore.State
+import com.egoriku.grodnoroads.shared.persistent.map.location.City
+import kotlinx.coroutines.flow.Flow
+
+@Stable
+interface CitySelectorComponent {
+
+    val state: Flow<State>
+
+    fun modify(preference: CitySelectorPref)
+    fun completeIntro()
+
+    @Stable
+    sealed interface CitySelectorPref {
+        data class DefaultCity(
+            val current: City = City.Grodno,
+            val values: List<City> = City.entries
+        ) : CitySelectorPref
+    }
+}
