+++ kmp/features/intro/src/commonMain/kotlin/com/egoriku/grodnoroads/intro/di/IntroModule.kt
@@ -1,0 +1,8 @@
+package com.egoriku.grodnoroads.intro.di
+
+import com.egoriku.grodnoroads.cityselector.di.citySelectorModule
+import org.koin.dsl.module
+
+val introModule = module {
+    includes(citySelectorModule)
+}
