+++ kmp/features/intro/src/commonMain/kotlin/com/egoriku/grodnoroads/onboarding/domain/component/OnboardingComponent.kt
@@ -1,0 +1,9 @@
+package com.egoriku.grodnoroads.onboarding.domain.component
+
+import androidx.compose.runtime.Stable
+
+@Stable
+interface OnboardingComponent {
+
+    fun completeOnboarding()
+}
