--- kmp/features/onboarding/src/commonMain/kotlin/com/egoriku/grodnoroads/onboarding/di/OnboardingModule.kt
@@ -1,13 +1,0 @@
-package com.egoriku.grodnoroads.onboarding.di
-
-import com.egoriku.grodnoroads.onboarding.domain.store.OnboardingStoreFactory
-import org.koin.dsl.module
-
-val onboardingModule = module {
-    factory {
-        OnboardingStoreFactory(
-            storeFactory = get(),
-            dataStore = get()
-        ).create()
-    }
-}
