--- kmp/features/onboarding/src/commonMain/kotlin/com/egoriku/grodnoroads/onboarding/ui/foundation/PageIndicator.kt
@@ -1,41 +1,0 @@
-package com.egoriku.grodnoroads.onboarding.ui.foundation
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.pager.PagerState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.unit.dp
-
-@Composable
-internal fun PageIndicator(
-    pagerState: PagerState,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier,
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(10.dp)
-    ) {
-        repeat(pagerState.pageCount) { index ->
-            val color = if (pagerState.currentPage == index) {
-                MaterialTheme.colorScheme.primary
-            } else {
-                MaterialTheme.colorScheme.outlineVariant
-            }
-            Box(
-                modifier = Modifier
-                    .clip(RoundedCornerShape(4.dp))
-                    .background(color)
-                    .size(16.dp, 4.dp)
-            )
-        }
-    }
-}
