+++ kmp/features/root/build.gradle.kts
@@ -40,7 +40,7 @@ kotlin {
             export(projects.kmp.features.eventReporting)
             export(projects.kmp.features.specialEventReminder)
             export(projects.kmp.features.mainflow)
-            export(projects.kmp.features.onboarding)
+            export(projects.kmp.features.intro)
             export(projects.kmp.features.settings.alerts)
             export(projects.kmp.features.settings.appearance)
             export(projects.kmp.features.settings.changelog)
@@ -65,7 +65,7 @@ kotlin {
             api(projects.kmp.features.specialEventReminder)
             api(projects.kmp.features.quickSettings)
             api(projects.kmp.features.mainflow)
-            api(projects.kmp.features.onboarding)
+            api(projects.kmp.features.intro)
             api(projects.kmp.features.settings.alerts)
             api(projects.kmp.features.settings.appearance)
             api(projects.kmp.features.settings.changelog)
