+++ kmp/features/root/src/commonMain/kotlin/com/egoriku/grodnoroads/root/di/KoinHelper.kt
@@ -5,7 +5,7 @@ import com.arkivanov.mvikotlin.main.store.DefaultStoreFactory
 import com.egoriku.grodnoroads.crashlytics.shared.di.crashlyticsModule
 import com.egoriku.grodnoroads.eventreporting.di.eventReportingModule
 import com.egoriku.grodnoroads.guidance.di.guidanceModule
-import com.egoriku.grodnoroads.onboarding.di.onboardingModule
+import com.egoriku.grodnoroads.intro.di.introModule
 import com.egoriku.grodnoroads.quicksettings.di.quickSettingsModule
 import com.egoriku.grodnoroads.settings.alerts.di.alertsModule
 import com.egoriku.grodnoroads.settings.appearance.di.appearanceModule
@@ -25,7 +25,7 @@ import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 fun appModule() = listOf(
-    onboardingModule,
+    introModule,
 
     guidanceModule,
     eventReportingModule,
